+++ src/evm_transition_tool/geth.py
@@ -67,8 +67,26 @@ class GethTransitionTool(TransitionTool):
             if not test_result["pass"]:
                 pytest.fail(f"Test failed: {test_result['name']}. Error: {test_result['error']}")
 
+    def get_blocktest_help(self) -> str:
+        """
+        Return the help string for the blocktest subcommand.
+        """
+        args = [str(self.binary), "blocktest", "--help"]
+        try:
+            result = subprocess.run(args, capture_output=True, text=True)
+        except subprocess.CalledProcessError as e:
+            raise Exception("evm process unexpectedly returned a non-zero status code: " f"{e}.")
+        except Exception as e:
+            raise Exception(f"Unexpected exception calling evm tool: {e}.")
+        return result.stdout
+
     def verify_fixture(
-        self, fixture_format: FixtureFormats, fixture_path: Path, debug_output_path: Optional[Path]
+        self,
+        fixture_format: FixtureFormats,
+        fixture_path: Path,
+        use_evm_single_test: bool,
+        fixture_name: Optional[str],
+        debug_output_path: Optional[Path],
     ):
         """
         Executes `evm [state|block]test` to verify the fixture at `fixture_path`.
@@ -87,6 +105,10 @@ class GethTransitionTool(TransitionTool):
         else:
             raise Exception(f"Invalid test fixture format: {fixture_format}")
 
+        if use_evm_single_test:
+            assert isinstance(fixture_name, str), "fixture_name must be a string"
+            command.append("--run")
+            command.append(fixture_name)
         command.append(str(fixture_path))
 
         result = subprocess.run(
@@ -100,7 +122,6 @@ class GethTransitionTool(TransitionTool):
 
         if debug_output_path:
             debug_fixture_path = debug_output_path / "fixtures.json"
-            shutil.copyfile(fixture_path, debug_fixture_path)
             # Use the local copy of the fixture in the debug directory
             verify_fixtures_call = " ".join(command[:-1]) + f" {debug_fixture_path}"
             verify_fixtures_script = textwrap.dedent(
@@ -119,9 +140,9 @@ class GethTransitionTool(TransitionTool):
                     "verify_fixtures.sh+x": verify_fixtures_script,
                 },
             )
+            shutil.copyfile(fixture_path, debug_fixture_path)
 
         if result.returncode != 0:
             raise Exception(
-                f"Failed to verify fixture via: '{' '.join(command)}'. "
-                f"Error: '{result.stderr.decode()}'"
+                f"EVM test failed.\n{' '.join(command)}\n\n Error:\n{result.stderr.decode()}"
             )