+++ src/pytest_plugins/test_filler/test_filler.py
@@ -5,6 +5,7 @@ Top-level pytest configuration file providing:
 and that modifies pytest hooks in order to fill test specs for all tests and
 writes the generated fixtures to file.
 """
+
 import warnings
 from pathlib import Path
 from typing import Generator, List, Optional, Type
@@ -183,7 +184,22 @@ def pytest_report_header(config, start_path):
         return
     binary_path = config.getoption("evm_bin")
     t8n = TransitionTool.from_binary_path(binary_path=binary_path)
-    return [f"{t8n.version()}, solc version {config.solc_version}"]
+    solc_version_string = Yul("", binary=config.getoption("solc_bin")).version()
+    return [f"{t8n.version()}, solc version {solc_version_string}"]
+
+
+def pytest_report_teststatus(report, config):
+    """
+    Disable test session progress report if we're writing the JSON fixtures to
+    stdout to be read by a consume command on stdin. I.e., don't write this
+    type of output to the console:
+
+    ```text
+    ...x...
+    ```
+    """
+    if config.getoption("output") == "stdout":
+        return report.outcome, "", report.outcome.upper()
 
 
 @pytest.fixture(autouse=True, scope="session")
@@ -240,7 +256,7 @@ def do_fixture_verification(request, t8n) -> bool:
 @pytest.fixture(autouse=True, scope="session")
 def evm_fixture_verification(
     request, do_fixture_verification: bool, evm_bin: Path, verify_fixtures_bin: Path
-) -> Optional[Generator[TransitionTool, None, None]]:
+) -> Generator[Optional[TransitionTool], None, None]:
     """
     Returns the configured evm binary for executing statetest and blocktest
     commands used to verify generated JSON fixtures.
@@ -296,6 +312,8 @@ def get_fixture_collection_scope(fixture_name, config):
 
     See: https://docs.pytest.org/en/stable/how-to/fixtures.html#dynamic-scope
     """
+    if config.getoption("output") == "stdout":
+        return "session"
     if config.getoption("single_fixture_per_file"):
         return "function"
     return "module"